% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgin.R
\name{dgin}
\alias{dgin}
\title{Title Density for the generalized inverse normal distribution}
\usage{
dgin(z, alpha, mu, tau, log = TRUE, quasi = FALSE)
}
\arguments{
\item{z}{quantile.}

\item{alpha}{degrees-of-freedom parameter.}

\item{mu}{similar to location parameter, controls asymmetry of the distribution.}

\item{tau}{similar to scale parameter, controls spread of the distribution.}

\item{log}{logical; should the log of the density be returned? Defaults to TRUE.}

\item{quasi}{logical; should the quasi-density value be returned? Defaults to FALSE.}
}
\value{
Numeric scalar with density.
}
\description{
Title Density for the generalized inverse normal distribution
}
\details{
Currently, only scalars are supported for the quantile and parameter values.
Density is supported on the entire real line, `z` and `mu` can take any value
in $(-\\infty, \\infty)$. Density is only defined for parameter values
`alpha`${} > 1$ or `tau`${} > 0$, so it is set to 0 outside of these values.
The quasi-density or kernel is the density without the normalization constant,
use `quasi = TRUE` for this behavior.
}
