% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgin1.R
\name{dgin1}
\alias{dgin1}
\title{Generalized inverse normal (log) density
Standardized density setting tau = 1
Internal function}
\usage{
dgin1(z, alpha, mu, log = TRUE, quasi = FALSE)
}
\arguments{
\item{z}{quantile}

\item{log}{boolean, optional(Should the log of the density be returned? Default is True.)}

\item{quasi}{boolean, optional(The quasi-density or kernel is the density without the normalization constant.
Should the quasi-density value be returned? Default is False.)}
}
\value{
dgin1
}
\description{
Generalized inverse normal (log) density
Standardized density setting tau = 1
Internal function
}
