% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtgin.R
\name{dtgin}
\alias{dtgin}
\title{Density for the generalized inverse normal distribution truncated to the positive or negative reals}
\usage{
dtgin(z, alpha, mu, tau, sign = TRUE, log = TRUE, quasi = FALSE)
}
\arguments{
\item{z}{real number (Distribution is supported on real line, z can be any real number.Currently, only scalar z is supported. Density is set to be 0 at z = 0.)}

\item{sign}{logical. `TRUE` implies truncation to positive numbers (`z` > 0)
and `FALSE` to negative numbers (`z` < 0)}

\item{log}{boolean, optional(Should the log of the density be returned? Default is True.)}

\item{quasi}{boolean, optional(The quasi-density or kernel is the density without the normalization constant.
Should the quasi-density value be returned? Default is False.)}
}
\value{
Numeric scalar with density.
}
\description{
Density for the generalized inverse normal distribution truncated to the positive or negative reals
}
\details{
Currently, only scalars are supported for the quantile and parameter values.
Density is supported on the positive reals (`z` > 0) when `sign = TRUE` and to
negative reals (`z` < 0) when `sign = FALSE`. `mu` can take any value
in $(-\\infty, \\infty)$. Density is only defined for parameter values
`alpha`${} > 1$ or `tau`${} > 0$, so it is set to 0 outside of these values.
The quasi-density or kernel is the density without the normalization constant,
use `quasi = TRUE` for this behavior.
}
