% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mshift.R
\name{mshift}
\alias{mshift}
\title{Internal function}
\usage{
mshift(z, mode, a, m, shift = TRUE)
}
\arguments{
\item{z}{2}

\item{mode}{3}

\item{a}{a degrees-of-freedom parameter}

\item{m}{3}

\item{shift}{3}
}
\value{
mshift
}
\description{
This is an internal function that adds two numbers.
}
