% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdam.R
\name{pbdam}
\alias{pbdam}
\title{Internal function}
\usage{
pbdam(a, z)
}
\arguments{
\item{a}{a degrees-of-freedom parameter}

\item{z}{A single numeric value.}
}
\value{
pbdam
}
\description{
This is an internal function that adds two numbers.
}
