% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgin.R
\name{rgin}
\alias{rgin}
\title{Generating random numbers from the generalized inverse normal distribution}
\usage{
rgin(size, alpha, mu, tau, algo = "hormann")
}
\arguments{
\item{size}{number of desired draws. Output is numpy vector of length equal to size)}

\item{algo}{string with desired algorithm to compute minimal bounding rectangle.
If "hormann", use the method from Hörmann and Leydold (2014). When "leydold", use the one from Leydold (2001).
Defaults to "hormann" and returns an error for any other values.}
}
\value{
Numeric vector of length `size`.
}
\description{
Generating random numbers from the generalized inverse normal distribution
}
\details{
Currently, only values of alpha > 2 are supported. For Bayesian posterior sampling,
alpha is always larger than 2 even for non-informative priors. The algorithm requires
calculating the probability of truncation region (either z < 0 or z > 0). It is more stable
to compute a probability bounded away from 0. As mu controls asymmetry,
when mu > 0, P(truncation region) = P(z>0) >= 50%, and this probability is computed.
If mu < 0, P(z<0) >= 50% and this region's probability is used.
}
