% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtgin.R
\name{rtgin}
\alias{rtgin}
\title{Generating random numbers from the generalized inverse normal distribution truncated to the positive or negative reals}
\usage{
rtgin(size, alpha, mu, tau, sign, algo, verbose = FALSE)
}
\arguments{
\item{size}{number of desired draws. Output is numpy vector of length equal to size)}

\item{sign}{logical. `TRUE` implies truncation to positive numbers (`z` > 0)
and `FALSE` to negative numbers (`z` < 0)}

\item{algo}{string with desired algorithm to compute minimal bounding rectangle.
If "hormann", use the method from Hörmann and Leydold (2014). When "leydold", use the one from Leydold (2001).
Defaults to "hormann" and returns an error for any other values.}

\item{verbose}{logical; should the acceptance rate from the ratio-of-uniforms
method be provided along with additional information? Defaults to FALSE.}
}
\value{
If `verbose = FALSE` (default), a numeric vector of length `size`.
Otherwise, a list with components `value`, `avg_arate`, and `ARiters`
}
\description{
Generating random numbers from the generalized inverse normal distribution truncated to the positive or negative reals
}
\details{
Currently, only values of alpha > 2 are supported. For Bayesian posterior sampling,
alpha is always larger than 2 even for non-informative priors.
Generate from positive region (`z` > 0) hen `sign = TRUE`, and from
negative region (`z` < 0) when `sign = FALSE`. When `verbose = TRUE`,
a list is returned containing the actual draw in `value`, as well as average
acceptance rate `avg_arate` and total number of acceptance-rejection steps `ARiters`.
}
