% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtgin1.R
\name{rtgin1}
\alias{rtgin1}
\title{Internal function}
\usage{
rtgin1(a, m, sign, algo = "hormann", verbose = FALSE)
}
\arguments{
\item{a}{a degrees-of-freedom parameter}

\item{m}{A single numeric value.}

\item{sign}{default}

\item{algo}{default TRUE}

\item{verbose}{default FALSE}
}
\value{
rtgin1
}
\description{
This is an internal function that adds two numbers.
}
